protocolutilsに重複する関数を集約
