+++ src/protocols/dem2SlopeProtocol.js
@@ -1,4 +1,5 @@
 import { addProtocol } from 'maplibre-gl';
+import { calculateTilePosition, getCalculateHeightFunction, calculatePixelResolution, calculateSlope } from './protocolUtils';
 
 function dem2SlopeProtocol(
     protocol = 'slope', 
@@ -26,8 +27,10 @@ function dem2SlopeProtocol(
                 tileY = parseInt(match[2], 10);
             }
 
+            const tileSize = 256; // タイルのサイズ（ピクセル）
+
             // 1ピクセルあたりの距離を計算
-            const pixelLength = calculatePixelLength(zoomLevel, tileY);
+            const pixelLength = calculatePixelResolution(tileSize, zoomLevel, tileY);
 
             // 周辺を含む9つのタイル画像のソース
             // index: 0 1 2
@@ -61,7 +64,6 @@ function dem2SlopeProtocol(
             ];
             }
 
-            const tileSize = 256; // タイルのサイズ（ピクセル）
             const buffer = 1; // 中央から読み込む範囲（ピクセル）
 
             // 結合用のキャンバスを作成
@@ -146,86 +148,3 @@ function dem2SlopeProtocol(
 }
 
 export { dem2SlopeProtocol };
-
-// タイルの位置を計算する関数
-function calculateTilePosition(index, tileSize, buffer) {
-    let sx, sy, sWidth, sHeight, dx, dy;
-    const col = index % 3;
-    const row = Math.floor(index / 3);
-
-    if (index === 4) { // 中央のタイル
-        return { sx: 0, sy: 0, sWidth: tileSize, sHeight: tileSize, dx: buffer, dy: buffer };
-    } else {
-        sx = (col === 0) ? tileSize - buffer : 0;
-        sWidth = (col === 1) ? tileSize : buffer;
-        dx = (col === 2) ? tileSize + buffer : col * buffer;
-
-        sy = (row === 0) ? tileSize - buffer : 0;
-        sHeight = (row === 1) ? tileSize : buffer;
-        dy = (row === 2) ? tileSize + buffer : row * buffer;
-
-        return { sx, sy, sWidth, sHeight, dx, dy };
-    }
-}
-
-// エンコーディングに応じた標高計算関数を返す関数
-function getCalculateHeightFunction(encoding) {
-    switch (encoding) {
-        case 'gsi':
-        case 'gsj':
-            return (r, g, b) => {
-                const x = r * 65536 + g * 256 + b;
-                const twoToThePowerOf23 = 8388608; // 2 ** 23
-                const twoToThePowerOf24 = 16777216; // 2 ** 24
-                if (x === twoToThePowerOf23) {
-                    return -99999;
-                }
-                return x < twoToThePowerOf23 ? 0.01 * x : 0.01 * (x - twoToThePowerOf24);
-            };
-        case 'mapbox':
-            return (r, g, b) => (r * 65536 + g * 256 + b) / 10 - 10000;
-        case 'terrarium':
-            return (r, g, b) => r * 256 + g + b / 256 - 32768;
-        default:
-            throw new Error('Unsupported encoding type');
-    }
-}
-
-// タイルのピクセル長さを計算する関数
-function calculatePixelLength(zoomLevel1, tileY1) {
-    // 度数法から弧度法へ変換
-    function toRadians(degrees) {
-        return degrees * (Math.PI / 180);
-    }
-
-    // 弧度法から度数法へ変換
-    function toDegrees(radians) {
-        return radians * (180 / Math.PI);
-    }
-
-    // タイルがカバーする緯度範囲を求める
-    const n = 2 ** zoomLevel1;
-    const latRad = Math.atan(Math.sinh(Math.PI * (1 - 2 * tileY1 / n)));
-    const latDeg = toDegrees(latRad);
-
-    // タイルがカバーする緯度の中心での1度あたりの距離（約111.32km/度、地球の半径を6371kmと仮定）
-    const latLengthPerDegree = 111.32 * 1000; // メートル単位
-
-    // ズームレベル7でのタイル1辺あたりの度数
-    const degreesPerTile = 360 / n;
-
-    // タイルの実際の長さを計算（緯度による補正を考慮）
-    const pixelLength = latLengthPerDegree * degreesPerTile * Math.cos(toRadians(latDeg)) / 256;
-
-    return pixelLength;
-}
-
-// 斜面を計算する関数
-// 産業技術総合研究所のシームレス傾斜量図の計算式を使用した。
-// https://gbank.gsj.jp/seamless/slope/
-function calculateSlope(H00, H01, H10, pixelLength) {
-    let dx = H00 - H01;
-    let dy = H00 - H10;
-    let slope = Math.atan(Math.sqrt(dx * dx + dy * dy) / pixelLength) * (180 / Math.PI);
-    return slope;
-}
