+++ src/protocols/weightFile.js
@@ -1,0 +1,58 @@
+// ガウシアン関数によるウェイトファイル
+export const weightFile25 = [
+    [0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000],
+    [0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.001, 0.001, 0.001, 0.001, 0.001, 0.001, 0.001, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000],
+    [0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.001, 0.001, 0.002, 0.002, 0.003, 0.004, 0.004, 0.004, 0.003, 0.002, 0.002, 0.001, 0.001, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000],
+    [0.000, 0.000, 0.000, 0.000, 0.000, 0.001, 0.002, 0.003, 0.005, 0.007, 0.009, 0.011, 0.011, 0.011, 0.009, 0.007, 0.005, 0.003, 0.002, 0.001, 0.000, 0.000, 0.000, 0.000, 0.000],
+    [0.000, 0.000, 0.000, 0.000, 0.001, 0.002, 0.004, 0.007, 0.012, 0.017, 0.023, 0.027, 0.029, 0.027, 0.023, 0.017, 0.012, 0.007, 0.004, 0.002, 0.001, 0.000, 0.000, 0.000, 0.000],
+    [0.000, 0.000, 0.000, 0.001, 0.002, 0.004, 0.009, 0.016, 0.027, 0.040, 0.053, 0.062, 0.066, 0.062, 0.053, 0.040, 0.027, 0.016, 0.009, 0.004, 0.002, 0.001, 0.000, 0.000, 0.000],
+    [0.000, 0.000, 0.001, 0.002, 0.004, 0.009, 0.018, 0.034, 0.056, 0.082, 0.108, 0.128, 0.135, 0.128, 0.108, 0.082, 0.056, 0.034, 0.018, 0.009, 0.004, 0.002, 0.001, 0.000, 0.000],
+    [0.000, 0.000, 0.001, 0.003, 0.007, 0.016, 0.034, 0.062, 0.103, 0.151, 0.200, 0.236, 0.249, 0.236, 0.200, 0.151, 0.103, 0.062, 0.034, 0.016, 0.007, 0.003, 0.001, 0.000, 0.000],
+    [0.000, 0.000, 0.002, 0.005, 0.012, 0.027, 0.056, 0.103, 0.169, 0.249, 0.329, 0.389, 0.411, 0.389, 0.329, 0.249, 0.169, 0.103, 0.056, 0.027, 0.012, 0.005, 0.002, 0.000, 0.000],
+    [0.000, 0.001, 0.002, 0.007, 0.017, 0.040, 0.082, 0.151, 0.249, 0.368, 0.486, 0.574, 0.607, 0.574, 0.486, 0.368, 0.249, 0.151, 0.082, 0.040, 0.017, 0.007, 0.002, 0.001, 0.000],
+    [0.000, 0.001, 0.003, 0.009, 0.023, 0.053, 0.108, 0.200, 0.329, 0.486, 0.641, 0.757, 0.801, 0.757, 0.641, 0.486, 0.329, 0.200, 0.108, 0.053, 0.023, 0.009, 0.003, 0.001, 0.000],
+    [0.000, 0.001, 0.004, 0.011, 0.027, 0.062, 0.128, 0.236, 0.389, 0.574, 0.757, 0.895, 0.946, 0.895, 0.757, 0.574, 0.389, 0.236, 0.128, 0.062, 0.027, 0.011, 0.004, 0.001, 0.000],
+    [0.000, 0.001, 0.004, 0.011, 0.029, 0.066, 0.135, 0.249, 0.411, 0.607, 0.801, 0.946, 1.000, 0.946, 0.801, 0.607, 0.411, 0.249, 0.135, 0.066, 0.029, 0.011, 0.004, 0.001, 0.000],
+    [0.000, 0.001, 0.004, 0.011, 0.027, 0.062, 0.128, 0.236, 0.389, 0.574, 0.757, 0.895, 0.946, 0.895, 0.757, 0.574, 0.389, 0.236, 0.128, 0.062, 0.027, 0.011, 0.004, 0.001, 0.000],
+    [0.000, 0.001, 0.003, 0.009, 0.023, 0.053, 0.108, 0.200, 0.329, 0.486, 0.641, 0.757, 0.801, 0.757, 0.641, 0.486, 0.329, 0.200, 0.108, 0.053, 0.023, 0.009, 0.003, 0.001, 0.000],
+    [0.000, 0.001, 0.002, 0.007, 0.017, 0.040, 0.082, 0.151, 0.249, 0.368, 0.486, 0.574, 0.607, 0.574, 0.486, 0.368, 0.249, 0.151, 0.082, 0.040, 0.017, 0.007, 0.002, 0.001, 0.000],
+    [0.000, 0.000, 0.002, 0.005, 0.012, 0.027, 0.056, 0.103, 0.169, 0.249, 0.329, 0.389, 0.411, 0.389, 0.329, 0.249, 0.169, 0.103, 0.056, 0.027, 0.012, 0.005, 0.002, 0.000, 0.000],
+    [0.000, 0.000, 0.001, 0.003, 0.007, 0.016, 0.034, 0.062, 0.103, 0.151, 0.200, 0.236, 0.249, 0.236, 0.200, 0.151, 0.103, 0.062, 0.034, 0.016, 0.007, 0.003, 0.001, 0.000, 0.000],
+    [0.000, 0.000, 0.001, 0.002, 0.004, 0.009, 0.018, 0.034, 0.056, 0.082, 0.108, 0.128, 0.135, 0.128, 0.108, 0.082, 0.056, 0.034, 0.018, 0.009, 0.004, 0.002, 0.001, 0.000, 0.000],
+    [0.000, 0.000, 0.000, 0.001, 0.002, 0.004, 0.009, 0.016, 0.027, 0.040, 0.053, 0.062, 0.066, 0.062, 0.053, 0.040, 0.027, 0.016, 0.009, 0.004, 0.002, 0.001, 0.000, 0.000, 0.000],
+    [0.000, 0.000, 0.000, 0.000, 0.001, 0.002, 0.004, 0.007, 0.012, 0.017, 0.023, 0.027, 0.029, 0.027, 0.023, 0.017, 0.012, 0.007, 0.004, 0.002, 0.001, 0.000, 0.000, 0.000, 0.000],
+    [0.000, 0.000, 0.000, 0.000, 0.000, 0.001, 0.002, 0.003, 0.005, 0.007, 0.009, 0.011, 0.011, 0.011, 0.009, 0.007, 0.005, 0.003, 0.002, 0.001, 0.000, 0.000, 0.000, 0.000, 0.000],
+    [0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.001, 0.001, 0.002, 0.002, 0.003, 0.004, 0.004, 0.004, 0.003, 0.002, 0.002, 0.001, 0.001, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000],
+    [0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.001, 0.001, 0.001, 0.001, 0.001, 0.001, 0.001, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000],
+    [0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000]
+];
+
+export const weightFile13 = [
+    [0.018, 0.034, 0.056, 0.082, 0.108, 0.128, 0.135, 0.128, 0.108, 0.082, 0.056, 0.034, 0.018], 
+    [0.034, 0.062, 0.103, 0.151, 0.200, 0.236, 0.249, 0.236, 0.200, 0.151, 0.103, 0.062, 0.034], 
+    [0.056, 0.103, 0.169, 0.249, 0.329, 0.389, 0.411, 0.389, 0.329, 0.249, 0.169, 0.103, 0.056], 
+    [0.082, 0.151, 0.249, 0.368, 0.486, 0.574, 0.607, 0.574, 0.486, 0.368, 0.249, 0.151, 0.082], 
+    [0.108, 0.200, 0.329, 0.486, 0.641, 0.757, 0.801, 0.757, 0.641, 0.486, 0.329, 0.200, 0.108], 
+    [0.128, 0.236, 0.389, 0.574, 0.757, 0.895, 0.946, 0.895, 0.757, 0.574, 0.389, 0.236, 0.128], 
+    [0.135, 0.249, 0.411, 0.607, 0.801, 0.946, 1.000, 0.946, 0.801, 0.607, 0.411, 0.249, 0.135], 
+    [0.128, 0.236, 0.389, 0.574, 0.757, 0.895, 0.946, 0.895, 0.757, 0.574, 0.389, 0.236, 0.128], 
+    [0.108, 0.200, 0.329, 0.486, 0.641, 0.757, 0.801, 0.757, 0.641, 0.486, 0.329, 0.200, 0.108], 
+    [0.082, 0.151, 0.249, 0.368, 0.486, 0.574, 0.607, 0.574, 0.486, 0.368, 0.249, 0.151, 0.082], 
+    [0.056, 0.103, 0.169, 0.249, 0.329, 0.389, 0.411, 0.389, 0.329, 0.249, 0.169, 0.103, 0.056], 
+    [0.034, 0.062, 0.103, 0.151, 0.200, 0.236, 0.249, 0.236, 0.200, 0.151, 0.103, 0.062, 0.034], 
+    [0.018, 0.034, 0.056, 0.082, 0.108, 0.128, 0.135, 0.128, 0.108, 0.082, 0.056, 0.034, 0.018], 
+];
+
+export const weightFile5 = [
+    [0.641, 0.757, 0.801, 0.757, 0.641],
+    [0.757, 0.895, 0.946, 0.895, 0.757],
+    [0.801, 0.946, 1.000, 0.946, 0.801],
+    [0.757, 0.895, 0.946, 0.895, 0.757],
+    [0.641, 0.757, 0.801, 0.757, 0.641],  
+];
+
+export const weightFile3 = [
+    [0.895, 0.946, 0.895],
+    [0.946, 1.000, 0.946],
+    [0.895, 0.946, 0.895],
+];
