+++ src/utils.js
@@ -290,7 +290,7 @@ function setLayerVisibility(map, layerId, visible) {
   }
 }
 
-// マジックナンバーの定数化
+// 定数
 const HILLSHADE_EXAGGERATION = 0.4;
 const HILLSHADE_HIGHLIGHT_COLOR = "rgb(255, 255, 255)";
 const CONTOUR_LINE_WIDTH_FACTOR = 2;
@@ -301,11 +301,11 @@ export function updateTerrainLayers(map, selectedDemSource, demSources, contourI
   // Update relief layer
   updateReliefLayer(map, selectedDemSource, demSources);
 
-  // Update slope layer
-  updateSlopeLayer(map, selectedDemSource, demSources);
-
   // Update cs layer
   updateCsLayer(map, selectedDemSource, demSources);
+  
+  // Update slope layer
+  updateSlopeLayer(map, selectedDemSource, demSources);
 
   // Update hillshade layer
   updateHillshadeLayer(map, selectedDemSource, demSources);
